test/f08: fix usage of mpi_type_hindexed and mpi_type_extent
Functions removed from MPI-1 does not exist in f08 interface. Replace
them with the replacement versions.

This is discovered due to accidental inclusion of old interface in the
f08 interface during development and the old interface have different
param types than the replacement versions.

The old tests worked because it simply falls back to linking with f77
interface.