python/c: refactor the mapping parameter
The parameter mapping is passed along only because a few lower level
function need it. Since the mapping concept is obscure at higher level,
having it in the interface makes the code obscrue. On the other hand, a
simpl string parameter "map_type" is more intuitive to understand. This
commit uses "map_type" or simply omits mapping whereever the mapping is
not directly needed.