comm: add split level "bindset" to HW_UNGUIDED split
This allows a proper split when processes are binded for example:

    mpirun -bind-to user:0+1,2+3,2+3 -n 3 ...

In this case, the processes are clearly grouped to two sets -- 1st set
bound to core 0 and 1, and 2nd set bound to core 2 and 3. There
applications doing such intende different sets run different tasks and
they would need to split the group based on the bindset. The hwloc
topology does not directly support this since neither set identifies a
single instance of hardware resource or they identify at different level
of topology binding. The "bindset" hint will allow user to split
according to the bindset.