ch4/ofi: make sure to use ctx[0]'s issued_cntr
When using Scalable context as vni context, it shares rma_cmpl_cntr with
context 0. While we can copy the rma_cmpl_cntr, we can't do that with
issued_cntr because it is not a pointer. So we need to use ctx[0]'s
issued cntr during `win_init_global` and `MPIDI_OFI_cntr_incr`.

NOTE: spread the logic of "USE_DOMAIN" and "SEP CTX" to multiple places
is not ideal.