test/mpi: add new hydra proc_binding test
Rewrite the test script using perl and write test configurations in a
single text file that is easy to generate and modify.

We run nearly the same set of tests as in the old test, removed those
that result in error, e.g. some topo.xml doesn't have sockets and we
don't support "board" type. In addition, we added some tests on `rr` and
`user` bind-to options.