op: refactor to use straight builtin op index
In ch4 RMA we played the trick of using MPI_OP_NULL is not needed but we
need a slot for compare_and_swap. Rather than shifting every builtin
index down by 1, simply give the 0th slot to compare_and_swap. This
makes the macro at MPIR-layer more intuitive to understand and more
robust as well.