hydra: Add parsing of MPIEXEC_PREFIX_DEFAULT
There was no handling of the environment variable MPIEXEC_PREFIX_DEFAULT in the
hydra code base. The variable adds the rank as a prefix to stdout in the form
[%r] by settings its value to anything greated than zero. A new function,
HYD_check_envs, was created to handle this and can be used for the handling of
other environment variables in the future. The already existing function,
prepend_rank_fn, checks for duplicate settings so using both MPIEXEC_PREFIX_DEFAULT
and -prepend-rank will correctly result in an error. These new settings were tested
manually and confirm that the changes work as intended.