hydra: Updated HYDU_set_str
HYDU_set_str was updated to take a const char* instead of a char*. No
where in the function is the char* value modified and this now allows
string literals to be passed. char* will be cast in to const char* and
there should be no issue with this change.