dataloop: fix builtin type alignment
The old code simply assumes the basic type alignment is the same as the
size. This is not true in principle. In fact, it is not true for complex
datatype and long double on 32-bit system. The fix is to use
MPIR_Datatype_builtintype_alignment (this is created at some to address
the struct alignment issue).

Fixes #3174