ch4: relax sender vci mapping under allow_overtaking hint
The comm hint mpi_assert_allow_overtaking removes the restrictions
to assign vci_idx to the sender threads such that message ordering
is maintained. This patch utilizes this opportunity for sender side
vni_idx mapping.

Relevant adjustment is done receiver side to deal with the relaxed
selection of transmit context by the sender.