typerep/yaksa: partial fix typerep_unpack_external
There is an obvious error in the wrong ordering in the unpack macros
that swapped the src and dest c types. This commit fixes that.

The packed/unpacked size in "external32" format are different from the
native size. This commit fixes the update of position using wrong size.

However, the code is still wrong in that it ignores the big-endian
requirement, unaligned data condition, and wrong packing of complex
types.

It also doesn't deal with datatypes that contain more than one basic
types, e.g. struct.