typerep/yaksa: fix MPIR_Typerep_pack/unpack_external
The old code doesn't meet the big-endian packing required by
"external32" datarep. This patch fixes it. It assumes the only datatype
potentially will have unequal size is 4/8-byte integers. It only
implements for this special case to keep the code simple and adds
assertions for all other cases. A potential missing is the long double
type -- left as FIXME.