typerep/yaksa: rewrite BASIC_convert into inline function
The previous code is correct. However, it raises compiler warnings since
the compiler may see code in `BASIC_convert64` for `int8_t`, even though
that branch should never be reached. Rewrite the code into inline
function and pass in the parameter as void pointer to avoid this
warning. The code is also more general. For example, we can easily
extend the code for more sizes and add support for non-aligned access.