ch4/ofi: remove lock from OFI call macro under lockless
For lockless thread safety model, we rely on the locking
from the OFI layer. So, no locking is needed at MPICH when
calling the ofi functions.

NOTE: there are still some OFI call macros that use locking,
e.g., MPIDI_OFI_VCI_CALL. Those macros are used by spawn and
are under non-critical paths, such as when creating a process.
It should be possible to get rid of those locks as well - a
possible area for future work.