mtest: add macros to manage dtp buffers
The possibility of testing device buffers adds complexity to the
original dtp tests. The miss-match of names and parameters has been a
constant source of bugs. This commit wraps the device buffer logic in a
set of macros to isolate the logic from the test itself and prevent
errors due to manual duplications.

It fixes errors where maxbufsize is used to init and check buffers
instead of using the actual dtp object size in
    test/mpi/rma/fence.c
    test/mpi/rma/accpscw1.c
    test/mpi/rma/putpscw1.c
    test/mpi/rma/epochtest.c
    test/mpi/rma/lock_x_dt.c