typerep: update FLOAT_convert to do simple byte swaps
The old code to do float conversion (for dataloop) is incorrect. Use
basic byte swap in stead.

Explain: the old conversion swaps the byte, then cast the result into
float and do the assignment. But since the swapped bytes may represent
an un-normalized floating point, the cast may not preserve the bits.