object: add MPIR_object_add_ref_unsafe
Some object reference updates do not need atomicity, for example, when
an object is created right before and is not shared with any other
threads.  The functionality of MPIR_object_add_ref_unsafe() is the same
as that of MPIR_object_add_ref(), while MPIR_object_add_ref_unsafe()
does not use a heavy fetch_add operation.