ch4/ofi: flush_send_queue on all vnis
We need flush send queues from all vnis. Rewrite the flushing with
non-blocking patterns to avoid potential deadlocks.

This fixes a hang on some tests with multi-vni, e.g. coll/allgatherv4.
It was broken since we switched from allreduce flushing to pt2pt
flushing.  It is not clear to me how the previous "allreduce" flushing
was able to clear the send queue on non-zero vnis