ch4: default MPIDI_CH4_MAX_VCIS to 64 when per-vci is on
Currently user have to configure with both --with-thread-cs=per-vci and
--with-ch4-max-vcis, as well as set MPIR_CVAR_CH4_NUM_VCIS to enable
multiple vcis. This is a bit cumbersome to remember and to explain. Let's
default `--with-ch4-max-vcis` to 64 -- the maximum number of request
pools -- to make it more convenient. It only adds 100K or so global
memory, which most applications won't be concerned with.