ch4: disable work queues under lockless
Lockless threading model does not need work queues. So, we remove
the need to queue during configuration and use in the code.

Note: Work queues are useful for handoff threading mode. Currently,
handoff support is turned off, so the work queues are not utilized.
Yet, the changes in this patch will be useful even after handoff is
made functional.