test: skip bool check in datatype/pack_external.c
Most C compilers will do boolean conversion on assignment, for example,
    bool a = 2;
here a will be set to 1 (true). Current implementation in MPICH, both
yaksa and dataloop, relies on compiler to do this boolean conversion.
However, in the case the compiler is treating bool as integers, it is
probably not critical to fix as long as the round-trip test still
passes. Thus we are skipping the check of pack buffer in the boolean
case.

This fixes (more like suppresses) the current test failure in Solaris
Studio.