ch4/ofi: add memory access ordering during probe
request->util_id is made atomic and release-acquire ordering
is applied when accessing it. This should prevent compiler
reordering of the memory accesses during probe operation.

During probe, we rely on util_id to notify a thread that the
the submitted OFI peek operation is completed by a progress
event. For this purpose, we maintain the memory access order
such that util_id is the last member of request object that
gets modified during progress and the first member that is
read from the probe routine after submitting the PEEK request.