ch4/ofi: refactor ofi provider code to init_provider.c
The provider selection code is very complex -- ~500 lines. Separate the
code so future work on it can be contained.

NOTE for rebasing --
This commit just moves code around plus adding necessary MPIDI_OFI_
prefix. This can be easily done manually. Once this commit is in place,
the rest of the commits will be much easier to cherry-pick.