ch4/ofi: prepost recv for disconnect
Prepost recv for disconnect messages from dynamic processes. This is
more robust and it will further allow disconnect to happen before
MPI_Finalize.

This is currently motivated by issues in verbs;ofi_rxm that it seems to
lose the messages if the sender goes away before the recv is posted.