ch4/ofi: flushing send queue by sending self-messages
Apparently sending self-messages can effectively flush sockets
provider's lightweight send queue. This is more efficient as it is a
local operation.

The fix is prompted by last commit that preposts recv for disconnect
messages. Flushing send queue was skipped when it is a single process,
but it may still need to flush for inter-world messages. It became an
issue due to pre-posting disconnect recv alters the timing.