ch4/ofi: force FI_MR_SCALABLE if mr_mode is 0
For libfabric >= v1.5, the FI_MR_SCALABLE is supposed to be the same as
mr_mode=0. However, with sockets provider, mr_mode=0 will return
FI_MR_BASIC. It is possible that the provider assume mr_mode=0 as no
hints, thus returning its default mr_mode. Setting it to FI_MR_SCALABLE
ensures the hints is honored.

NOTE: this is possible that sockets provider specific behavior.