ch4/ofi: use the required version in capability sets
The behavior of fi_getinfo is determined by the requested interface
version, rather than the actual library version, which makes sense as it
is only way for backward compatibility. Thus we need consistently set
hints based on the requested interface version.

Each capability set reflects the version that we set, thus we need use
the preset rather than override it with current library version.