sched: prepare to be used for perisistent collective
Add MPIR_Sched_free and MPIR_Sched_reset to support using sched in
persistent collectives. MPIR_Sched_start interface is modified not to
use a pointer to MPIR_Sched_t, thus the ownership is not explicitly
transferred.

The current design allow sched callbacks to alter entries, and making it
impossible to reset. Fortunately, we only need it in nonblocking
contextid algorithm, which does not need be persistent -- hopefully,
never will. To make the code more clear, we differentiate the kind into
regular, persistent, and generalized. Assertions are added where we
don't expect the entries to be changed.

In order to use in persistent collective, we need associate the state
buffer pointer with the sched. MPIR_Sched_alloc_state is added to opaquely
allocate and retrieve this buffer. It will be automatically freed when
sched is freed.

Sometime the same nonblocking operation may be composed of multiple
stages that each require temporary buffers. Allow allocating multiple
buffers to a single schedule. The maximum buffers is hard coded to
keep the code simple. If there are algorithms require more buffers,
simply increase the constant.