comm: Add default value to info hints
When an info hint is initialized, it's impossible to tell the difference
between a default value of `0` and a user-specified value of `0`. If the
user actually wanted `0` and MPICH overwrites it because it thinks the
user didn't specify, that's a problem.

Add a default value for each info hint so the implementation can tell
the difference.