config: Capture hwloc dependencies for static builds
Hwloc dependencies were lost when we converted to configuring the
library externally rather than use the supplied macros. This broken
static builds, since they would fail to link all the needed
libraries. This patch depends on a patch to the hwloc build system to
create the hwloc.pc file in the embedded configuration.