ch4/ofi: reduce av table dimensions
In principle, we need to store mappings of every remote context in every
local context. However, if we are careful and control the insertion
order, we can result in identical address mapping in each local context.
While this incurs extra code logic and makes code more fragile -- for
example if we ever need to implement a partial world initialization,
it probably will break -- it saves the storage of the local av table as
well as potentially simplifying the av lookup process. For large-scale
scenarios this saving can be significant enough to justify the
additional code complexity.