misc: use collective fallbacks during init
The communicator during initialization may not have been fully
initialized. For example, the csel is not initialized and the auto
algorithm can not be used. We define a set of algorithms as the fallback
alternatives and use them during init.

This commit fixes the crash when MPIR_CVAR_CH4_ROOTS_ONLY_PMI is on.