ch4/ofi: Add hint for NIC preference
Adds a new hint to indicate the NIC that the user wants a particular
communicator to use for all traffic. If this value is set at any
process in a communicator, it must be set at all processes in the
communicator. It also cannot be changed while there is outstanding
communication.