init: expose MPIR_init_lock
The creation of builtin comms are still part of the initializations that
is potentially being delayed in a session model. It need protection from
race condition using the same init lock. This commit exposes the
init_lock from static to an extern symbol so it can be used later in
MPIR_Comm_create_from_group_impl.