romio fp_ind and type index traversal within a view
Part of a file's state in romio is "fp_ind" which appears to mean the location for the next byte to be written.  And a key operation is translating that byte offset into a location within the file's view datatype, eg a count of how many file_type extents + which index within the flattened file view type corresponds to a write that would touch that byte.

The below gist
https://gist.github.com/markalle/4f0ff6b028769cc8f03dea9b881d5543
creates a datatype with extent=12 that's `[int, space, space]` but also sets LB=4.  This results in a flattened file_type of
> flat_file->indices[0] = 4  and ->blocklens[0] = 0
> flat_file->indices[1] = 0  and ->blocklens[1] = 4
> flat_file->indices[2] = 16  and ->blocklens[2] = 0

Then it writes 2 MPI_INT, followed by 2 more MPI_INT.  In this case the fp_ind is meaningful for both calls: first call fp_ind==0 and second call fp_ind==24.  But if you go to the first occurrence of
```
            n_filetypes = (offset - flat_file->indices[0]) / filetype_extent;
```
inside ADIOI_NFS_WriteStrided, the LB marker messes up the n_filetypes calculation (it decides we've walked 1 extent instead of 2 into the file_view).  And from there the bookkeeping goes into the weeds and it doesn't write the data to the right place.

I made a utility function to do roughly the same thing as the code I deleted in this commit, but more paranoid of strange edge cases.

The other problem with fp_ind (but I don't have a gist testcase for it) is that sometimes there's nothing stopping fp_ind from being updated to the next flat->indices[] entry for a LB/UB marker, which can be arbitrary offsets which makes the above lookup function impossible to guarantee a unique result.

My current code tries to work within that design and allows the lookup to identify a LB/UB marker, although I think that makes fp_ind not well defined and it 'should' only be allowed to point to data indices.  But fp_ind is used so many places it would be a large task to try reviewing them all.

Signed-off-by: Mark Allen <markalle@us.ibm.com>