Add cvar to dump a corefile on abort
By default, MPICH either exits or calls PMI when aborting. Add a cvar
to allow the user to force MPICH to call abort() and generate a
corefile, which is useful for development/debugging. This patch adds
the abort() call inside MPID_Abort, since that is what MPICH calls
internally when an error is detected.