ch4/ofi: add provider preference
We currently select providers based on a matching score, which measures
the provider capability against a default capability set and a minimal
capability set. When the matching score is the same, we used to pick the
first provider returned by libfabric, which often ranks based on its
assumed performance. However, sometimes, that preference does not reflect
the best choice for MPICH. For example, with the currently embedded
libfabric, it picks `UDP;ofi_rxd` over `sockets` for tcp only system,
when `UDP;ofi_rxd` does not pass our full test suite.

The correct solution is probably to identify the missing check and
enhance our default capability so that the matching score should better
reflect conformance. In the meantime, we can simply fix the selection by
hard code a preference score.