maint: detect duplicated cvar definitions
We may accidentally or intentionally define the same cvars in multiple
locations. This adds the check for duplications.

Sometime this can be intentional. For example,
`MPIR_CVAR_OFI_USE_PROVIDER` can be used in both ch3:ofi or ch4:ofi, and
it is preferable to define in both places so in the future removing one
wouldn't break another.