ch4/ofi: duplicate MPIR_CVAR_OFI_USE_PROVIDER definition
MPIR_CVAR_OFI_USE_PROVIDER is currently defined in
src/mpid/ch3/channels/nemesis/netmod/ofi/ofi_init.c. This can cause
confusion as it is also used in ch4:ofi but the definition can't be
found there. In addition, if we ever decide to remove ch3:ofi, it will
break ch4:ofi, and it is undesirable.

Duplicate the cvar definition in ch4:ofi.