config: fix PAC_APPEND_FLAG and PAC_PREPEND_FLAG
The usage `$FGREP -e "\<$1\>"` is wrong because fgrep doesn't accept
regex. This commit replace it with `$FGREP -e "$1"`. It has the risk of
false positive of matching partial string. We can be hopeful that such
case never occur with our usages. If it does occur, it can be locally
worked around by either swapping appending order or simply replace with

    $FLAG="$FLAG $1"