sched: remove racing condition after completion
As soon as we call MPIR_Request_free, another thread may start another
round of persist coll and racing with the next line `s->req = NULL`.

Even both are setting `s->req` to NULL, the racing codnition corrupts
the data.

The setting of `s->req` to NULL is unnecessary here.