ch4/part: Convert nm/shm init calls to hooks
Directly call generic partitioned initialization calls from the ch4
layer. Lower-level APIs are converted to hooks to allow for
optimization. The intention here is to always use the generic
initialization for request matching, since it is off the fast path and
does not require any additional resources. Netmods and shmmods can
hook into initialization to perform additional work for optimization.