romio/daos: add support for DAOS pool and container labels
DAOS now supports labels in addition to uuids for pools and container.
Update the DAOS ADIO driver to work with both labels and uuids,
depending on the API version of DAOS being used and adjust the environment
variables to be able to set a prefix for the direct UNS path.