ch4: remove the usage of am_isendv
The only place we use am_isendv is when we need send short flattened
datatype within the am header. It adds complexity but brings little
benefit. Most lower-layers do is to allocate buffer and memcpy anyway.

This commit composes the am header in the upper layer, thus remove the
need of am_isendv.