ch4/ucx: consolidate am_isend_reply
The current optimization -- skipping an extra memcpy of large data --
was only applied in MPIDI_NM_am_isend_reply in ucx, utilizing the
assumption that large data is only sent in the reply segment. This
assumption is unnecessary and entangles the logic in both the upper
layer and lower layer. This patch consolidates am_isend_reply with
am_isend.