romio gpfs: free of NULL/uninitialized address, found by hdf5
This resembles a previous fix that I had only applied to the
others_req[] variable, but my_req[] is made similarly and also
needed the initialization of the .offsets field and a check for
non-NULL when freeing it.

I didn't whittle down the testcase to reproduce it, I just
observed the failure in a full hdf5 run.

Signed-off-by: Mark Allen <markalle@us.ibm.com>