test: fix impls/mpich/comm/comm_dup.c
Fix a typo that resulted in checking the original comm instead of the
duplicated comm.

The test is testing that info hints should not be copied with
MPI_Comm_dup. However, it is okay for communicator to have some built-in
hints with default values. Update the test to test it properly.