ch4: fix release_gather in reduce
We can't use the recvbuf for local gather when the local rank is not
root. Allocate temporary buffer in this case.

Note: the use of temporary buffer probably has a negative effect on the
algorithm performance. Hopefully, a good collective tuning/selection
will avoid picking the algorithm when it has worse performance.