ch4: remove MPIDI_NM_av_is_local netmod api
The av_is_local query is used at ch4-layer to dispatch between netmod
api and shm api. It is not up to netmod to decide the locality. Locality
need be consistent at the ch4-layer. In the case when
MPIDI_CH4_DIRECT_NETMOD is enabled or locality is not available, we
should always assume is_local is false.