mpir: Assert when request is not allocated
There are two options for how to handle a request not being allocated.

1. Assert when the request is NULL

This isn't ideal as it's pretty bad error handling. On the other hand,
the alternative is to...

2. Change every call to allocate a request to check for successful
output and perform error handling.

This is obviously more correct, but has a performance impact. In the
debug case, we already check to see if things were allocated correctly.