ch4: dynamically grow av tables and remove free list
Instead of initializing with an arbitrary number of entries, use a
simple scheme to dynamically grow the av table.

The benefit of managing the free list is not clear. For one, it is just
a pointer table, space is not critical even if we only grow and never
free. For two, iterating a list of thousands of entries to find a free
entry in the non-critical path of connecting new dynamic processes
should be okay. Removing this mechanism makes the code easier to
maitnain. We add a simple replacement mechanism of reusing free entries.

Now that we simplified the avt manager, the usage of
MPIDIU_get_max_n_avts need be replaced by MPIDIU_get_n_avts.