ch4/shm: remove debug prints in topotree.c
The code in topotree is very complex. The flurry of debugging statements
made the code reading even more difficult. While the debug prints helps
when one already has a mental understanding and all the debug prints
makes sense to him/her. However, if one need to understand the code
first, the debug output won't make sense anyway, and makes the
comprehension more difficult. Most developers won't retain their
understanding after a while, so we need assume that trying understand
the code is more often at priority.

Debug prints are easy to insert. The time of inserting the code compared
to the debug session is negligent.