ch4/ofi: Confirm provider NIC is up
Some providers choose to report the NIC status in the struct returned by
fi_getinfo rather than filtering out NICs that are down before returning
them. In this case, we need to check the status of the link state in the
nic part of the provider struct before we use it.