ch4/ipc: clean up get_map_device
Local function get_map_device decides which device to map the remote
memory on. The code was not clear to read. This patch cleans it up.

* Raname recv_type and recv_dev_id. Since it is not just used for
send/recv case, and even for send/recv case, the context of recv is
not clear at the address mapping time, thus we removed the recv from the
names.

* Pass in local device id rather than MPL_gpu_device_handle_t. Since the
remote is passed in a device id, the local part should also be an id to
be more intuitive.

* Re-arrange the logic. We either map to local device or remote device.
The branch structure should make this clear.

* Add TODO comment on the non-contig case. There must be a performance
concern that decides the mapping choice for the non-contig case. This is
not clear and need to be explained. Besides, when local buffer is host
memory, map the remote memory arbitrary to device 0 is questionable.