ch4/ofi: replace MPIDI_OFI_progress_do_queue
The current cq buffering system does not work with multiple VCI. But
before we fix it, let's test whether we need the buffering system at
all. The original recursive locking concerns no longer exist. While it
is nice that we prevent recursive progress situations, but it should
work in principle.