ch4/ofi: further refactor cq buffering routines
It makes the code difficult to read when we can read a single statement,
or even a single variable name in one line. Here we repeat the long name
pattern in nearly every line, and we are about to make it even longer by
moving the variable into per-vci array. Use local macros to shorten
them.