ch3/nemesis: use stronger atomics
It has been reported failures using ch3 on ARM Cortex-A75, issue #5309.
It was confirmed this patch fixes the issue. It remain to be fully
analyzed how the fiaures happen and whether this patch is over zealous,
mostly due to lack of testing systems. Meanwhile, I think we should
apply the patch now rather than leave the code broken.