ch4/part: Change ready counter to count down instead of up
Set the ready counter to the number of partitions and count down more
like a traditional completion counter. This allows us to remove a
redundant assertion that the counter is within the allowable range. We
can also use the incomplete output argument from MPIR_cc_decr to inform
other code logic.