ch4: add vci parameters to MPIDIG_request_create
The request need be created using the right vci pool and store the vci
information in the request so later am_send_reply can use the correct
vcis.

The src_vci and dst_vci can be confusing, especially for protocols with
back and forth exchanges. Use local_vci and remote_vci instead.