ch4: refactor ch4-layer per_vci global
Simply setting minimum struct size to cacheline is not sufficient to
prevent false sharing. For example, the last a few fields may share the
same cacheline as the first a few fields of the next array element if
the totoal struct size is greater than but not multiple of the cacheline
size. Use MPL_ATTR_ALIGNED instead.

Note, neither setting the alignment to the whole struct (which modifies
the alignment when the struct itself is the member of another struct),
nor setting the alignment to the array (which modifies the alignment of
the entire array rather than individual elements) would work. But
setting the alignment of the struct member, including its flexible array
member, will work.

Also rename vci to per_vci as in per vci globals. `vci` is used as
integer value elsewhere, and can be quite confusing here to name a
struct.