ch4/ipc: initialize sockets for fd-based ipc handles
For a set of node-local process, we setup sockets between local rank 0
and all other local ranks. Normally, this socket would be used to
communicate the fd-based ipc handles between to processes.

We do not setup sockets between every pair of local ranks. This is
because the strategy is to implement a "shareable" handle for fd-based
ipc. The sockets will be used to share device file descriptors in order
to transfer fd-based ipc handles without the use of UNIX sockets.