ch4/ofi: make sure am message buffers are consecutive
When fi_sendv is not available, if the whole message fits in the header
buffer space, we can directly send it as a single buffer.

When the message doesn't fit in the header space, we'll have to allocate
a separate send buffer. We can reuse the pack_buffer pointer since we
won't need both.

Move the header pool macros to ofi_pre.h since its context is relevant
there.