maint: Avoid parsing warnings in submodules
The clmake script parses out all of the warnings and other unexpected
output when compiling. However, it's not realistic to expect all
dependencies to always be warning clean. This ends up causing problems
that MPICH shouldn't have to resolve.

So every time clmake detects that its in a submodule (anything with a
'modules' prefix), it turns itself off until it comes out of the
submodule.