gpu/ipc: initialization for shareable fd ipc handles
Local rank 0 opens file descriptors for each local device and shares
them with all other local processes.

This mechanism assumes the devices are opened in the same order as level
zero. It has been requested that level zero provides an API for opening
these device handles in order to guarantee they are opened in the same
order.

Co-authored-by: Gengbin Zheng <gengbin.zheng@intel.com>