mpl/ze: use device fds for ipc management
Create and map ipc handles using the shareable device fds.

During ipc handle creation, make it to a shareable ipc handle by
converting from the dma_buf fd to a GEM handle. This handle and the
source device are then stored in the shared device fd to be accessed on
the remote side during ipc mapping.

When mapping an ipc handle, convert the shareable ipc handle (GEM handle
+ source device) back to a dma_buf fd and open the handle. When p2p is
supported, the handle is opened on the destination device. When p2p is
not supported, the handle is opened on the source device. This is so
that Yaksa treats it as a cross-device copy and takes the appropriate
path.

Co-authored-by: Gengbin Zheng <gengbin.zheng@intel.com>