ch4/ofi: fix lmt_cntr mechanism
The purpose of lmt_cntr is to track which rdma read is the last read so
its completion can proceed to next step. However, we know whether the
read message is the last read at the time we issue fi_read, so we could
directly set the boolean flag rather than using the perisistent field in
the original rreq object.