mpl/ze: implement ipc using pidfd_getfd syscall
Level zero uses file descriptors to implement ipc handles. To prevent
the need for UNIX sockets or for creating device fds, use pidfd_getfd to
duplicate opened fds in order to support ipc in the ze backend. This
implementation requires Linux Kernel 5.6.0 or higher. This requirement
is checked at runtime. If it is not met, ipc is disabled for the upper
layer.