ch4/ofi: add barrier during finalize for psm2
This can help prevent a situation where a rank may try to connect
to an OFI endpoint that may have already been closed by a remote
rank during finalize.

One scenario where this can happen is: a rank waiting to receive
a message whereas the sender rank has sent the data and moves
forward to the finalize step.