ch4/ofi: send huge ctrl header directly
In the old mechanism, we send a portion of the data natively and send a
separate control header. The mechanism requires synchronization between
the two parts.

Instead, let's send the control header directly as a native send. Upon
receiving, we can check whether it is, in fact, a huge message, and act
accordingly.