ch4/ofi: simplify the control structure
Separate the huge info header from the control structure since we no
longer use it as a control message.

Make control a union since the current control handler is frameworked as
a general handler and thus fit better with a union structure.

Remove ofi_control.h since it is the same as am_send_header.