ch4/ofi: huge send to use sreq handle in cq_data
Previously we embed msgsize in cq_data so that probe can have the
msgsize without extra query. However, in the case that sender sends a
huge message and receiver use a tiny buffer that is too small to receive
the control info, we won't be able to ack sender since we will not have
sender request information.

The patch put the sreq handle in the cq data in stead. Consequently, we
have to run extra control message for probe in order to get the msgsize.