ch4/ofi: fix huge message data_sz
Since now we no longer send actual data in the first message, we no
longer need to modify the receive buffer size to handle the data
underflow issue.

The data_sz stored in the rreq is used to detect data truncation
problem, which I don't think the previous code handles.