ch4/ofi: revamp huge message handling
Clean up the data structure to be more specific.

Differentiate probe and mprobe. The former can be discarded when the
control isn't ready. If we don't discard unsuccessful probe and put it in
a queue, it can cause issues when another probe or recv come to
interfere. Mprobe, on the other hand, is guaranteed to match once, so
there is no issue.

Persist the remote info with the original rreq. This avoids the use of
separate hash maps to look up. It is also cleaner to track.