datatype: always free pairtypes at finalize
Always free pairtypes regardless of reference count at finalize to prevent
handle leaks.

Reference count of predefined types can be non-zero if the reference
count is obtained via MPI_Type_get_contents. The standard asks user not
to free a pre-defined type from MPI_Type_get_contents. However, this is
practically difficult for user to follow since it is not always easy to
tell whether a datatype is predefined or not. Thus, reference count the
predefined datatype probably is more robust.

Regardless, we should not leak handles after finalize.