ch4/ofi: remove MPIDI_OFI_ENABLE_SENDV
This commit removes the option to use fi_sendv in the am path.
The most significant difference is in the extra copy of short am
message. Since it is only a small memcpy, it is not clear how
significant is the benefit. Remove to simplify the code for now. We
always can add it back by reverting this patch.