gpu/ipc: change visible_dev_global_id to int8
This variable is a flag so it does not need to be an int. This change is
necessary due to the addition of subdevice support in ze. For systems
with 6 or more Intel GPUs with tiles, the visible_dev_global_id would be
larger than a cache line size and fail due to size restrictions on
MPIDU_Init_shm_put/get.