ch4: fix MPIDIG_do_put
In the old code, we set target_count to either the actual count or
data_sz, and then adjust if it is actually a flattened datatype. This is
very confusing. In fact, it misses the case when the flattened type is
sent within the header and the callback neglected to adjust the count.

In stead, we should set target_count individually in each branch. In the
case we are sending flattened datatype, the target_count should be the
original count, thus avoiding latter adjusting the count in callbacks.