mpl: Add streaming copy
The stream copy function is used when requested by the caller. The
benefit of this function is that it does not bring the target memory
into cache when doing a write. The is particularly important when
writing data from one socket to memory physically located in space
attached to another socket.