fortran: add string util functions in mpi_fortimpl.h
Rather than generate the string conversion functions, add them in
mpi_fortimpl.h.

Even though we are going to replace the buildiface script, it is updated
in the commit to use the new string functions, so we can test it.

Instead of initiating MPI_F_ARGV_NULL with empty string, it is more
reliable to check for the address explicitly, the same way as we do for
all other builtin names.