f08: accept sizes for fint/aint/count/cint
A simple aint_is_int is not sufficient to check whether a POLY functions
actually need different interfaces unless we check the actual integer
sizes. Pass in the actual type sizes instead. This is necessary when
e.g. INTEGER is 8-byte rather than 4-byte.

When Fint is not the same as c_int, we can't have functions passing Fint
into a c_int parameter even in a `c_int == kind(0)` branch. The
compiler, e.g. gfortran, will complain and refuse to compile. Thus, we
also need option for c_int size and skip generation of one branch or the
other.