bsend: Hardcode MPI_BSEND_OVERHEAD value
The MPICH ABI initially defined this value to 96, which was based on the
size of MPII_Bsend_data_t. However the value was accidentally changed
(96->88 on x86_64) when the MPII_Bsend_data_t structure was
modified. This could cause a one-way incompatibility where an
application using the new, lower value may crash if run with a library
using the old, higher value.

To address this, we hardcode the value to 96, which is an upper bound
for the user. As long as the size of MPII_Bsend_data_t stays at or below
96 bytes, then there will be no ABI compatibility issue. A check is
added in configure ensure the struct size meets the ABI requirement.