bsend: Add dummy var to MPII_Bsend_data to restore ABI
[efb8d24a9] removed a field from this structure, which in turn modified
the value of MPI_BSEND_OVERHEAD and the MPICH ABI. This was
unintended. Add a dummy variable to the struct to restore the
ABI. Long-term, we may want to hardcode an upper limit for
MPI_BSEND_OVERHEAD for various architectures, and use a configure check
to ensure this structure fits within that limit.

This discovery was made with the help of ABI Compliance
Checker. https://github.com/lvc/abi-compliance-checker.