ch4/ofi: fix bug in get_huge_complete
We need save the comm_ptr before calling MPIDI_OFI_recv_event because it
may free the request pointer and make rreq->comm inaccessible.

This is especially true when the receive is an any source receive, when
the user visible request is the shm request, and the netmod request will
get freed after copied over.

Fixes pmodels/mpich#5607.