ch4/am: Use union to differentiate send/recv request fields
It is sometimes hard to tell how fields in generic requests are being
used, for example in what type of operation. Use a union to add context
about the type of operation. In this case, send and recv requests. RMA
requests also use these fields based on the type (or stage) of
operation.