coll: refactor iallreduce_tsp_recursive_exchange_common
The step1 sends data from non-participating ranks -- extra ranks in
excess of power-of-k -- to participating ranks; step3 does the reverse.
Both steps are quite orthoganol to the step 2, which is the actual
recursive exchange algorithm.

This commit refactors the interface to make both step1 and step3 more
independent of the actual step 2 algorithm. In addition, it also
abstracts the temporary buffer allocation since it is also need be used
during step 1 and step 3.

It also add a routine that user can call to re-order the peer list in
case of non-commutative reductions. It will be used in the next
refactoring.